Release awarded credits/evos slowly over time.
Credits/evos are not awarded right away, but released slowly over time.
When a player dies, the credit not yet released are lost. This motivates
people to stay alive, in particular, not to feed by trying to get
evos/credits by desperate suicide attacks.

The rate is controlled by variable *g_creditReleaseTime*. It holds the
number of seconds it takes to release 2000 credits or 9 evos.