Bounce marks and sounds for culled particles.
Before a particle with "bounce cull" could not display marks and make a
sound when hitting something. I moved the code (very simply) so that
when a particle has "bounce cull" but also a bounce sound / bounce mark,
the mark/sound is processed before the particle dies. (So it's rather a
hit mark/sound.) This allows for things like a drop of liquid hitting
something, splatting and turning into a mark etc.